functions/requirements.txt
firebase-functions==0.4.2
firebase-admin>=6.2.0
slack-bolt>=1.18.0
slack-sdk>=3.21.3
omegaconf>=2.3.0
python-dateutil>=2.8.2
pytz>=2023.3
aiohttp>=3.8.5
functions-framework>=3.4.0
flask>=2.3.3
python-dotenv>=1.0.0